Create Kraken.Daemon.Internal.
        For non-exported Daemon components.
